quick rename to remove duplicate reference
