+++ src/ECER.Clients.RegistryPortal/ecer.clients.registryportal.client/src/config/character-reference-evaluation-form.ts
